change column type to float

+++ database/migrations/2020_06_23_00522_alter_availability_perc_column.php
@@ -14,7 +14,7 @@ class AlterAvailabilityPercColumn extends Migration
     public function up()
     {
         Schema::table('availability', function (Blueprint $table) {
-            $table->double('availability_perc', 9, 6)->default(0.000000)->change();
+            $table->float('availability_perc', 9, 6)->default(0.000000)->change();
         });
     }
     /**
